+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 - `qflags` code prunning matching processing v2.
 - added `config.db_schemas.platforms_schemas.get_schema_as_dict`
 - improved `config.db_schemas.platforms_schemas` by adding `version` parameters
+- Added `sftp_tools.get_local_filepaths_from_sftp_downloaded_files_as_dict`
+- Updated
 
 ## [v0.13.0]
 - Implemented data processing V2.0 with new phencam QFLAGs and L2 and L3 derived products with weights.